+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The following is a curated list of changes in the Enact agate module, newest cha
 - `agate/LabeledIconButton` max-width so that huge sized icon will not be cut off
 - `agate/Spinner` to pause the animation when `paused` prop is true
 - `agate/Picker` picker item width in horizontal for silicon
+- `agate/SliderButton` to not show console error in qa-sampler
 
 ## [1.0.0] - 2020-10-14